requirements.txt
# Installation: pip install -r requirements.txt

# FastAPI and server (0.103.1)
fastapi>=0.115.2  
uvicorn>=0.23.2
python-multipart>=0.0.6

# Database ORM
sqlalchemy>=2.0.21

# Authentication and security
passlib>=1.7.4
bcrypt==3.2.2  # Downgraded to a version compatible with passlib
python-jose>=3.3.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0

# Rate Limiting
slowapi

# Testing
pytest>=7.0.0
httpx>=0.24.0  # Required for TestClient

# Utilities
email-validator>=2.0.0.post2
